cache: move transfer to objects
Moving it as is for now.
