+++ dvc/cache/base.py
@@ -1,6 +1,5 @@
 import itertools
 import logging
-from concurrent import futures
 from concurrent.futures import ThreadPoolExecutor
 from copy import copy
 from typing import Optional
@@ -8,14 +7,13 @@ from typing import Optional
 from funcy import decorator
 from shortuuid import uuid
 
-from dvc.dir_info import DirInfo
 from dvc.exceptions import CacheLinkError, DvcException
 from dvc.progress import Tqdm
 from dvc.remote.slow_link_detection import (  # type: ignore[attr-defined]
     slow_link_guard,
 )
 
-from ..objects import HashFile, ObjectFormatError, Tree
+from ..objects import HashFile, ObjectFormatError
 
 logger = logging.getLogger(__name__)
 
@@ -128,84 +126,6 @@ class CloudCache:
         callback = kwargs.get("download_callback")
         if callback:
             callback(1)
-
-    def _transfer_file(self, from_tree, from_info):
-        from dvc.utils import tmp_fname
-        from dvc.utils.stream import HashedStreamReader
-
-        tmp_info = self.tree.path_info / tmp_fname()
-        with from_tree.open(
-            from_info, mode="rb", chunk_size=from_tree.CHUNK_SIZE
-        ) as stream:
-            stream_reader = HashedStreamReader(stream)
-            # Since we don't know the hash beforehand, we'll
-            # upload it to a temporary location and then move
-            # it.
-            self.tree.upload_fobj(
-                stream_reader,
-                tmp_info,
-                total=from_tree.getsize(from_info),
-                desc=from_info.name,
-            )
-
-        hash_info = stream_reader.hash_info
-        return tmp_info, hash_info
-
-    def _transfer_directory_contents(self, from_tree, from_info, jobs, pbar):
-        rel_path_infos = {}
-        from_infos = from_tree.walk_files(from_info)
-
-        def create_tasks(executor, amount):
-            for entry_info in itertools.islice(from_infos, amount):
-                pbar.total += 1
-                task = executor.submit(
-                    pbar.wrap_fn(self._transfer_file), from_tree, entry_info
-                )
-                rel_path_infos[task] = entry_info.relative_to(from_info)
-                yield task
-
-        pbar.total = 0
-        with ThreadPoolExecutor(max_workers=jobs) as executor:
-            tasks = set(create_tasks(executor, jobs * 5))
-
-            while tasks:
-                done, tasks = futures.wait(
-                    tasks, return_when=futures.FIRST_COMPLETED
-                )
-                tasks.update(create_tasks(executor, len(done)))
-                for task in done:
-                    yield rel_path_infos.pop(task), task.result()
-
-    def _transfer_directory(
-        self, from_tree, from_info, jobs, no_progress_bar=False
-    ):
-        dir_info = DirInfo()
-
-        with Tqdm(total=1, unit="Files", disable=no_progress_bar) as pbar:
-            for (
-                entry_info,
-                (entry_tmp_info, entry_hash),
-            ) in self._transfer_directory_contents(
-                from_tree, from_info, jobs, pbar
-            ):
-                self.add(entry_tmp_info, self.tree, entry_hash)
-                dir_info.trie[entry_info.parts] = entry_hash
-
-        return Tree.save_dir_info(self, dir_info)
-
-    def transfer(self, from_tree, from_info, jobs=None, no_progress_bar=False):
-        jobs = jobs or min((from_tree.jobs, self.tree.jobs))
-
-        if from_tree.isdir(from_info):
-            return self._transfer_directory(
-                from_tree,
-                from_info,
-                jobs=jobs,
-                no_progress_bar=no_progress_bar,
-            )
-        tmp_info, hash_info = self._transfer_file(from_tree, from_info)
-        self.add(tmp_info, self.tree, hash_info)
-        return hash_info
 
     def cache_is_copy(self, path_info):
         """Checks whether cache uses copies."""
