+++ dvc/remote/base.py
@@ -512,8 +512,14 @@ class Remote:
         return ret
 
     def transfer(self, from_tree, from_info, jobs=None, no_progress_bar=False):
-        return self.cache.transfer(
-            from_tree, from_info, jobs=jobs, no_progress_bar=no_progress_bar
+        from dvc.objects import transfer
+
+        return transfer(
+            self.cache,
+            from_tree,
+            from_info,
+            jobs=jobs,
+            no_progress_bar=no_progress_bar,
         )
 
     @staticmethod
